Move jobbers into jobber.py, add shell jobber
